+++ game/Assets/Script/LoadScript.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: cd0cf8eb271844b4c808bdef1907c243
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
